requirements.txt
flask
playwright
google-auth
google-api-python-client
gunicorn
